update new address ui with stardust route
