+++ client/src/app/components/stardust/AssetProps.tsx
