+++ client/src/app/components/stardust/NFT.scss
