+++ client/src/app/components/stardust/NFTProps.tsx
