+++ client/src/app/routes/stardust/Addr.scss
@@ -155,6 +155,87 @@
     }
   }
 
+  .asset-cards {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+
+    .section--assets {
+      background: #F8F9FA;
+      border-radius: 6px;
+      width: 469px;
+      height: 85px;
+  
+      .inner--asset {
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        margin: 30px;
+  
+        .assets {
+          display: flex;
+          flex-direction: column;
+        }
+      }
+  
+      .svg-navigation {
+        width: 16px;
+        height: 16px;
+        margin-top: 6px;
+      }
+
+      @include tablet-down {
+        width: 350px;
+      }
+    }
+
+    .section--NFT {
+      background: #F8F9FA;
+      border-radius: 6px;
+      width: 469px;
+      height: 85px;
+      margin-left: 20px;
+  
+      .inner--asset {
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        margin: 30px;
+  
+        .assets {
+          display: flex;
+          flex-direction: column;
+        }
+      }
+  
+      .svg-navigation {
+        width: 16px;
+        height: 16px;
+        margin-top: 6px;
+      }
+
+      @include tablet-down {
+        margin-top: 20px;
+        width: 350px;
+        margin-left: 0px;
+      }
+    }
+
+    @include tablet-down {
+      flex-direction: column;
+    }
+  }
+
+  .nft--section {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+  }
+
+  .no-border {
+    border-bottom: none;
+  }
+
   .transaction--section {
     @include tablet-down {
       .section--header {
